refactor proxy aggregator heap and ring buffer manager
- extract active-bucket heap logic into proxy_active_bucket_heap
- move worker ring buffer ownership and lookup into ring_buffer_mgr
- pre-bind each proxy bucket to its target ring buffer during init
- make ring_buffer_mgr get path a read-only indexed lookup after initialization
- decouple supernode worker ring buffer binding from proxy_aggregator
- remove redundant input_rbs storage from supernode state
- align proxyConfig field types with proxy_aggregator_config
- remove thin proxy_batch_bucket helper wrappers in favor of direct field access
- inline flush idle sleep policy and drop flush_scheduler_idle_sleep_us
- add Chinese documentation for ring buffer mapping, bucket-to-rb binding, proxy thread model, and flush timing
- update proxy-related unit tests to match the new heap and ring buffer manager structure

+++ src/server.h
@@ -1866,9 +1866,9 @@ typedef enum childInfoType {
 typedef struct hotkeyStats hotkeyStats;
 
 typedef struct proxyConfig {
-    int batch_limit;              /* 0 means default */
-    int time_limit_us;            /* 0 means default */
-    int max_supernodes;           /* 0 means default */
+    size_t batch_limit;           /* 0 means default */
+    uint64_t time_limit_us;       /* 0 means default */
+    size_t max_supernodes;        /* 0 means default */
 } proxyConfig;
 
 struct redisServer {
